Add a class to collect the reference from Godot
This uses the Godot GDScript language server to dump the class reference as
JSON. To use this tool, you need to copy the file into your project and run it
from the Godot editor: it's currently not possible to run it from the command
line with Godot's `--script` option.

Known issue at the time of commit: the tool won't parse the class's docstring.

Closes #12
Closes #11
Closes #10
Closes #9
Closes #1